feature/86cw10vv5: created function for getting service data
